Merge Develop

+++ frappe/desk/doctype/desktop_icon/desktop_icon.py
@@ -9,8 +9,8 @@ from frappe import _
 import json
 import random
 from frappe.model.document import Document
-from six import iteritems
-
+from six import iteritems, string_types
+from frappe.utils.user import UserPermissions
 
 class DesktopIcon(Document):
 	def validate(self):
@@ -71,7 +71,7 @@ def get_desktop_icons(user=None):
 				if standard_icon.blocked:
 					icon.hidden = 1
 
-					# flag for modules_setup page
+					# flag for modules_select dialog
 					icon.hidden_in_standard = 1
 
 				elif standard_icon.force_show:
@@ -95,7 +95,7 @@ def get_desktop_icons(user=None):
 				icon.hidden = 1
 
 		# sort by idx
-		user_icons.sort(lambda a, b: 1 if a.idx > b.idx else -1)
+		user_icons.sort(key = lambda a: a.idx)
 
 		# translate
 		for d in user_icons:
@@ -127,12 +127,14 @@ def add_user_icon(_doctype, _report=None, label=None, link=None, type='link', st
 			frappe.session.user)[0][0] or \
 			frappe.db.sql('select count(*) from `tabDesktop Icon` where standard=1')[0][0]
 
-		module = frappe.db.get_value('DocType', _doctype, 'module')
-		module_icon = frappe.get_value('Desktop Icon', {'standard':1, 'module_name':module},
-			['name', 'icon', 'color', 'reverse'], as_dict=True)
-
 		if not frappe.db.get_value("Report", _report):
 			_report = None
+			userdefined_icon = frappe.db.get_value('DocType', _doctype, ['icon','color','module'], as_dict=True)
+		else:
+			userdefined_icon = frappe.db.get_value('Report', _report, ['icon','color','module'], as_dict=True)
+
+		module_icon = frappe.get_value('Desktop Icon', {'standard':1, 'module_name':userdefined_icon.module},
+			['name', 'icon', 'color', 'reverse'], as_dict=True)
 
 		if not module_icon:
 			module_icon = frappe._dict()
@@ -149,8 +151,8 @@ def add_user_icon(_doctype, _report=None, label=None, link=None, type='link', st
 				'type': type,
 				'_doctype': _doctype,
 				'_report': _report,
-				'icon': module_icon.icon,
-				'color': module_icon.color,
+				'icon': userdefined_icon.icon or module_icon.icon,
+				'color': userdefined_icon.color or module_icon.color,
 				'reverse': module_icon.reverse,
 				'idx': idx + 1,
 				'custom': 1,
@@ -171,7 +173,7 @@ def add_user_icon(_doctype, _report=None, label=None, link=None, type='link', st
 @frappe.whitelist()
 def set_order(new_order, user=None):
 	'''set new order by duplicating user icons (if user is set) or set global order'''
-	if isinstance(new_order, basestring):
+	if isinstance(new_order, string_types):
 		new_order = json.loads(new_order)
 	for i, module_name in enumerate(new_order):
 		if module_name not in ('Explore',):
@@ -196,11 +198,13 @@ def set_desktop_icons(visible_list, ignore_duplicate=True):
 	if the desktop icon does not exist and the name is a DocType, then will create
 	an icon for the doctype'''
 
-	# clear all custom
-	frappe.db.sql('delete from `tabDesktop Icon` where standard=0')
+	# clear all custom only if setup is not complete
+	if not int(frappe.defaults.get_defaults().setup_complete or 0):
+		frappe.db.sql('delete from `tabDesktop Icon` where standard=0')
 
-	# set all as blocked
-	frappe.db.sql('update `tabDesktop Icon` set blocked=0, hidden=1')
+	# set standard as blocked and hidden if setting first active domain
+	if not frappe.flags.keep_desktop_icons:
+		frappe.db.sql('update `tabDesktop Icon` set blocked=0, hidden=1 where standard=1')
 
 	# set as visible if present, or add icon
 	for module_name in visible_list:
@@ -228,7 +232,7 @@ def set_hidden_list(hidden_list, user=None):
 	'''Sets property `hidden`=1 in **Desktop Icon** for given user.
 	If user is None then it will set global values.
 	It will also set the rest of the icons as shown (`hidden` = 0)'''
-	if isinstance(hidden_list, basestring):
+	if isinstance(hidden_list, string_types):
 		hidden_list = json.loads(hidden_list)
 
 	# set as hidden
@@ -354,6 +358,69 @@ def sync_from_app(app):
 
 	return modules_list
 
+@frappe.whitelist()
+def update_icons(hidden_list, user=None):
+	"""update modules"""
+	if not user:
+		frappe.only_for('System Manager')
+
+	set_hidden_list(hidden_list, user)
+	frappe.msgprint(frappe._('Updated'), indicator='green', title=_('Success'), alert=True)
+
+def get_context(context):
+	context.icons = get_user_icons(frappe.session.user)
+	context.user = frappe.session.user
+
+	if 'System Manager' in frappe.get_roles():
+		context.users = frappe.db.get_all('User', filters={'user_type': 'System User', 'enabled': 1},
+			fields = ['name', 'first_name', 'last_name'])
+
+@frappe.whitelist()
+def get_module_icons(user=None):
+	if user != frappe.session.user:
+		frappe.only_for('System Manager')
+
+	if not user:
+		icons = frappe.db.get_all('Desktop Icon',
+			fields='*', filters={'standard': 1}, order_by='idx')
+	else:
+		frappe.cache().hdel('desktop_icons', user)
+		icons = get_user_icons(user)
+
+	for icon in icons:
+		icon.value = frappe.db.escape(_(icon.label or icon.module_name))
+
+	return {'icons': icons, 'user': user}
+
+def get_user_icons(user):
+	'''Get user icons for module setup page'''
+	user_perms = UserPermissions(user)
+	user_perms.build_permissions()
+
+	from frappe.boot import get_allowed_pages
+	allowed_pages = get_allowed_pages()
+
+	icons = []
+	for icon in get_desktop_icons(user):
+		add = True
+		if icon.hidden_in_standard:
+			add = False
+
+		if not icon.custom:
+			if icon.module_name=='Learn':
+				pass
+
+			elif icon.type=="page" and icon.link not in allowed_pages:
+				add = False
+
+			elif icon.type=="module" and icon.module_name not in user_perms.allow_modules:
+				add = False
+
+		if add:
+			icons.append(icon)
+
+	return icons
+
 palette = (
 	('#FFC4C4',),
 	('#FFE8CD',),
@@ -404,3 +471,16 @@ palette = (
 	('#4F8EA8', 1),
 	('#428B46', 1)
 )
+
+@frappe.whitelist()
+def hide(name, user = None):
+	if not user:
+		user = frappe.session.user
+
+	try:
+		set_hidden(name, user, hidden = 1)
+		clear_desktop_icons_cache()
+	except Exception:
+		return False
+
+	return True
